Updated Spawner Logic and Enemy Gravity
Updated Spawner Logic to fix spawn time and spawn delay when the game is restarted. Problem was due to using Time.time to trigger speed increases. But Time.time is not reset when the scene is reloaded. Fixed by using Time.timeSinceLastLoad. Also, modified the enemy gravity from +0.2 per 10 seconds to +0.5 per 20 seconds, may still tweak. Work for tomorrow is finishing up the spawner logic between respawns and maybe tweaking the high and low ranges for the random values.