Changed Spawner modification from random to longest spawn time first
Instead of selecting random spawner every 3 seconds to decrease the spawn time of, now choose the spawner with the largest spawn time for a more consistent decrease of overall spawntime. If the game goes long enough the blocks will be falling every 0.25 seconds with 5x gravity.

Done with spawner logic for now until playtesting Saturday.

Will try to focus on CameraController or Score placement tomorrow and friday because the screen scaling seems off